refactor: 침대, 침실, 욕실 갯수 타입 Integer -> Double 로 변경

+++ be/src/main/java/kr/codesquad/airbnb/domain/Room.java
@@ -23,9 +23,9 @@ public class Room {
 
     private Integer pricePerNight;
     private Integer maxNumberOfGuest;
-    private Integer bedroom;
-    private Integer bed;
-    private Integer bathroom;
+    private Double bedroom;
+    private Double bed;
+    private Double bathroom;
 
     @OneToMany(mappedBy = "room")
     private List<RoomAmenity> roomAmenities = new ArrayList<>();
